kindex: Particle System emitter camera smooth OpenGL engine comments
git-svn-id: svn://localhost/steel.new/trunk@103 99e027b2-380c-0410-a737-a5784e1a7bde

+++ steel/trunk/code/game/objects/path.cpp
@@ -57,7 +57,7 @@ void TagPath::process(steel::time curTime, steel::time frameLength, PhysicEngine
 	{
 		GameObj *n = parent->findChildren(target[(currentTarget + 1)%cnt].id);
 		if(!n) return;
-		coord k = (curTime - smoothStart) / smoothLen;
+		coord k = (float)((curTime - smoothStart) / smoothLen);
 
 		if(k>=1)
 		{
